[RPS-56] Run acc. tests against standalone Hippo instance
Through the use of Maven profiles 'cargo.run', 'acceptance-test' and
'acceptance-test-hippo', it is now possible to run automated acceptance
tests against an instance of Hippo CMS that was either started manually,
running constantly in the background, or one that is started and
stopped automatically prior and after the tests, respectively.

The details of how to use the profiles have been described in the
readme.