[RPS-170] RPS-171 Refactor properties to beans
Moved a number of get methods from DatasetComponent and
PublicationComponent to the relevent bean.
This is more object oriented and I've changed the freemarker templates
to look at the property directly rather than inserting as a seperate
variable.
Also added PowerMockito to the project as I needed to mock a static
method to simulate queries.