[RPS-204] CI migration: make folder names human-friendly
EximImport was re-set to its original version and MigratorImporterScript
was added as THE import script to use to import content generated by
the Migrator application; readme has been updated accordingly.

This is to reflect the fact that MigratorImporterScript is tailored
specifically to consume content generated by the Migrator rather than
more generic output of EximExport.

'FOLDER' JSON files generated by the Migrator application are now being
processed by the import script and so the folders are now created with
display names as specified by property 'jcr:localizedName' in files
generated by the Migrator.

It is now also possible to create empty folders upon migration.