[RPS-192] Migrate CI Compendium
The Migrator application now includes in the migrated content
Compendium.

Confluence page 'Structure Mapping' contains Excel spreadsheet, whose
third sheet is used as the source of mapping. Note that while the
structure of this file is expected to not change, its content is a work
in progress and we expect to get updated version(s) of it in the near
future as CI team completes the mapping.

Also, as per the same Confluence page:

Titles of Publications and Series are now generated correctly
accross the board (not just Compendium), that is, they are now set to
the names of appropriate 'ancestor' folders (with appropriate
'Archived' prefixes, where applicable).

Folders immediately under 'root' Compendium folder are now prefixed
with 'Compendium'.