[RPS-220] Set Freemarker template MIME type
FreeMarker recommend using ftlh and ftlx file extensions to turn on HTML
and XML auto-escaping (along with a "incompatible_improvements" setting
which was introduced in version 2.3.24). All template files renamed to
use the .ftlh extension. Removed legacy ?html escaping from templates.

To check the status of FreeMarker configuration, add the following to
a  template:
<p>Output format: ${.output_format}
<p>Auto-escaping: ${.auto_esc?c}