[RPS-216] Page not found page
Changed the page not found page so it uses the base layout so the
styling is the same as the rest of the site. Also set the HTTP status
code to 404.

Added to the sitemap so if you enter an invalid URL you will be
directed to the new 404 page.

Also cleaned up some items in the sitemap that either weren't used
or were specified in the wrong place.