[RPS-253] Add accessibility help page and namespace cleanup
Enabled a populated version of accessibility help document
to be bootstraped in the development module, with a blank
document being bootstraped in other environments on first
deployment for the user to populate and maintain (similar
to taxonomy and terms and conditions). For subsequent
deployments, as the document will already exist, the user
maintained version will not be overwritten.

Moved About document type from publicationsystem namespace
to common namespace.  Change involved updating config to
allow document types under the common namespace to also
be included as part of the resource location filters, which
enable automatic scanning for content-bean annotated classes.