[RPS-308] CI migration: better remote import
It is now possible for the import process executing on the servers to
obtain the content to be imported automatically from external resource.

This eliminates the need for current, largely manual, time consuming
and error-prone method of perorming the import (import locally, manually
export the entire content in batches as XML, manually upload/import each
batch individually on the server).

From now on, the Migrator application generates a single
'import-package.zip' archive file, containing all that's needed for
import - both the EXIM JSON files and attachments.

MigratorImportHelper class was introduced to encapsulate all server-side
logic of downloading, unzipping and post-processing of the content,
making it ready for import.

The post processing is about adjusting the paths to attachment files
embedded in EXIM JSON files to match actual paths on the server: EXIM
appears to require that these paths are absolute and so they need to
be adjusted between environments.