[RPS-337] CI landing pages add repeating sections
Add new compound type to allow for repeating text/resources sections
on CI landing page. This will allow constuction of content to fit
in with requirements for the SHMI landing page.  Within this repeating
section is a Rich Text Editor field, which is restricted to only allow
basic formatting (bold, italic, underline) and numbered or bullet
point lists.