[RPS-454] Added test for preview
Added a test for the preview that checks if a document can be created
and previewed. In order for this to run against the production only
config I have created a new module called local that only contains
the host.yaml config that allows the site to be mounted on localhost.