[RPS-450] RPS-450 cap S3 concurrent uploads
Configuration of SchedulingS3Connector service can be changed via
Console on the fly, without restarting the server. This stops the thread
pools from accepting new tasks, un-registers the previous instance of
the service, initialises new thread pools and registers new instance
of the service.

S3ConnectorValve used to only retrieve an instance of this service once,
upon system start and would not use the latest instance of the service,
which would lead to the stopped thread pools reject download requests,
breaking the Channel Manager Preview in the process.

This fix eliminates this problem by making the valve retrieve the latest
instance of the service for each download request rather than get it
once and store as a field for the lifetime of the valve object.