[RPS-487] NPE in SearchTabsComponent
It appears that when the 'timeframe' facet is specified last
in the search URL and its value is the only one in the query that makes
it not match any documents, the resultSet returned from the
facetNavigationBean is null rather than empty. The null value was the
source of the offending exception seen in the logs.

There is now a null-check in place to prevent the exception from being
thrown.