[RPS-365] Create breadcrumbs for CI pages
Added breadcrumbs for CI content (series,publication,
datasets). If the document sits under one of the
clinical indicator folders (i.e. CCG, SHMI), then a
breadcrumb back to that CI landing page is also added
to the hierarchy.

Note: 2 separate breadcrumbs had to be created, one
which uses the old styling (breadcrumb-ci-old-styling)
and one which uses the new styling (breadcrumb-ci). Once
the dataset template has been updated with the new styling,
then breadcrumb-ci-old-styling can be removed, and replaced
with the same breadcrumb that the publication/series page
uses.