[RPS-532] Implement High Charts
Added the ability to add high chart sections to publication pages
The data is parsed from an uploaded spreadsheet and converted into
a JSON string which is fed into the highcharts libary to render the
data as an interactive visualisation.

Also fixed a bug with the index for publications with 1 page.