[RPS-534] Validate charts data file
Custom validator triggered on document save now verifies that the
provided input for Highcharts is valid by letting the parser try to
interpret the parameters (including the input file) and reporting
failure in a controlled manner if the parsing fails with exception.

Also, derived data function that used to trigger the conversion has now
been replaced by workflow task kicking in on document save and updating
document draft variant.