[RPS-497] RPS-450 S3 logs fix and upload cap values
Concurrently modified counters are now designated 'volatile' to prevent
the risk that different threads getting different values of the fields
in the shared instance.

Also, caps on the maximum numbers of concurrent downloads and uploads
via CMS have now been set to values concluded in the S3 integration
performance test.