[DW-751] Add option to suppress table sorting
Checkbox option added to CKeditor Table component to suppress
table sorting (which is enabled by default). Note: implementation
involves duplicating and modifying the file:

table > dialogs > table.js

Therefore, as with all CMS customisations, there is
a risk when upgrading the Hippo version of this
functionality breaking (CKEditor table or CKEditor in general)
and will need to be checked.