[DW-803] Add relevance expressional inference engine add-on
Also fixes Channel manager event list and service list
components so they are rendered correctly.

Inference engine characteristics, and characteristic UI plugins
which are added need to be bootstraped into the content (config
as code), otherwise Hippo clears any existing values on each deployment.

Includes an example of an Expressional rule, with a couple of goals
for news and event.

Update news template with a component on the right pane to
demo displaying of relevance functionality.