[ADZ-55] API Specifications - Try it now
Hides the 'Servers' drop-down, meaning that
only Sandbox server will be supported by
the feature for now.

This will likely change once the support
for authentication will be scrutinised more
closely in a separate story.

For now, support for just Sandbox puts us
on the same level as Apigee which is enough
for MVP.