[ADZ-1551] Last Change Check Instant fix
Adds dedicated node to track metadata used
during import of API Specifications, moving
the website:lastChangeCheckInstant to it.

This is so that the timestamp is not tracked
on the API Specification documents themselves,
as updating it there used to unnecessarily save
the entire documents on each batch job run,
even if the documents themselves did not change.